+++ network-mux/CHANGELOG.md
@@ -4,6 +4,16 @@
 
 ### Breaking changes
 
+* Removed `Netowrk.Mux.Compat` module with legacy API.
+* `Ouroboros.Network.Mux.toApplication` was removed. 
+* `Ouroboros.Network.Mux.mkMiniProtocolBundle` was renamed to
+  `mkMiniProtocolInfos`, its type changed.
+* Removed `MiniProtocolBundle` newtype wrapper.
+* Generalised `Channel` type and provide `ByteChannel` type alias.
+* Provide additional APIs in the `Network.Mux.Channel` for creating channels
+  and byte channels.
+* `MuxBearer` has a `name` field.
+
 ### Non-breaking changes
 
 * Fix compilation with `tracetcpinfo` flag.